requirements.txt
Package                  Version
------------------------ ------------
absl-py                  2.1.0
aiohttp                  3.9.5
aiosignal                1.3.1
annotated-types          0.6.0
anyascii                 0.3.2
asttokens                2.4.1
attrs                    23.2.0
audioread                3.0.1
Babel                    2.15.0
bangla                   0.0.2
blinker                  1.8.2
blis                     0.7.11
bnnumerizer              0.0.2
bnunicodenormalizer      0.1.6
cachetools               5.3.3
catalogue                2.0.10
certifi                  2024.2.2
cffi                     1.16.0
charset-normalizer       3.3.2
click                    8.1.7
cloudpathlib             0.16.0
comm                     0.2.2
confection               0.1.4
contourpy                1.2.1
coqpit                   0.0.17
cycler                   0.12.1
cymem                    2.0.8
Cython                   3.0.10
dateparser               1.1.8
debugpy                  1.6.7
decorator                4.4.2
docopt                   0.6.2
einops                   0.8.0
encodec                  0.1.1
exceptiongroup           1.2.0
executing                2.0.1
ffprobe                  0.5
filelock                 3.14.0
Flask                    3.0.3
fonttools                4.51.0
frozenlist               1.4.1
fsspec                   2024.5.0
g2pkk                    0.1.2
google-api-core          2.19.0
google-api-python-client 2.129.0
google-auth              2.29.0
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.0
googleapis-common-protos 1.63.0
grpcio                   1.63.0
gruut                    2.2.3
gruut-ipa                0.13.0
gruut_lang_de            2.0.0
gruut_lang_en            2.0.0
gruut_lang_es            2.0.0
gruut_lang_fr            2.0.2
gTTS                     2.5.1
hangul-romanize          0.1.0
httplib2                 0.22.0
huggingface-hub          0.23.0
idna                     3.7
imageio                  2.34.1
imageio-ffmpeg           0.4.9
importlib_metadata       7.1.0
inflect                  7.2.1
ipykernel                6.29.3
ipython                  8.24.0
itsdangerous             2.2.0
jamo                     0.4.1
jedi                     0.19.1
jieba                    0.42.1
Jinja2                   3.1.4
joblib                   1.4.2
jsonlines                1.2.0
jupyter_client           8.6.1
jupyter_core             5.7.2
kiwisolver               1.4.5
langcodes                3.4.0
language_data            1.2.0
lazy_loader              0.4
librosa                  0.10.2.post1
llvmlite                 0.42.0
lxml                     5.2.2
marisa-trie              1.1.1
Markdown                 3.6
MarkupSafe               2.1.5
matplotlib               3.9.0
matplotlib-inline        0.1.7
more-itertools           10.2.0
moviepy                  1.0.3
mpg123                   0.4
mpmath                   1.3.0
msgpack                  1.0.8
multidict                6.0.5
murmurhash               1.0.10
nest_asyncio             1.6.0
networkx                 2.8.8
nltk                     3.8.1
num2words                0.5.13
numba                    0.59.1
numpy                    1.26.4
nvidia-cublas-cu12       12.1.3.1
nvidia-cuda-cupti-cu12   12.1.105
nvidia-cuda-nvrtc-cu12   12.1.105
nvidia-cuda-runtime-cu12 12.1.105
nvidia-cudnn-cu12        8.9.2.26
nvidia-cufft-cu12        11.0.2.54
nvidia-curand-cu12       10.3.2.106
nvidia-cusolver-cu12     11.4.5.107
nvidia-cusparse-cu12     12.1.0.106
nvidia-nccl-cu12         2.20.5
nvidia-nvjitlink-cu12    12.4.127
nvidia-nvtx-cu12         12.1.105
oauth2client             4.1.3
oauthlib                 3.2.2
openai-whisper           20231117
opencv-python            4.9.0.80
packaging                24.0
pandas                   1.5.3
parso                    0.8.4
pexpect                  4.9.0
pickleshare              0.7.5
pillow                   10.3.0
pip                      24.0
platformdirs             4.2.1
pooch                    1.8.1
preshed                  3.0.9
proglog                  0.1.10
prompt-toolkit           3.0.42
proto-plus               1.23.0
protobuf                 4.25.3
psutil                   5.9.8
ptyprocess               0.7.0
pure-eval                0.2.2
pyasn1                   0.6.0
pyasn1_modules           0.4.0
pycparser                2.22
pydantic                 2.7.1
pydantic_core            2.18.2
pydub                    0.25.1
Pygments                 2.18.0
pynndescent              0.5.12
pyparsing                3.1.2
pypinyin                 0.51.0
pysbd                    0.3.4
python-crfsuite          0.9.10
python-dateutil          2.9.0
pytrends                 4.9.2
pyttsx3                  2.90
pytz                     2024.1
PyYAML                   6.0.1
pyzmq                    25.1.2
regex                    2024.5.15
requests                 2.31.0
requests-oauthlib        2.0.0
rsa                      4.9
safetensors              0.4.3
scikit-learn             1.4.2
scipy                    1.13.0
setuptools               69.5.1
six                      1.16.0
smart-open               6.4.0
soundfile                0.12.1
soxr                     0.3.7
spacy                    3.7.4
spacy-legacy             3.0.12
spacy-loggers            1.0.5
srsly                    2.4.8
srt                      3.5.3
stack-data               0.6.2
SudachiDict-core         20240409
SudachiPy                0.6.8
sympy                    1.12
tensorboard              2.16.2
tensorboard-data-server  0.7.2
thinc                    8.2.3
threadpoolctl            3.5.0
tiktoken                 0.7.0
tokenizers               0.19.1
torch                    2.3.0
torchaudio               2.3.0
tornado                  6.4
tqdm                     4.66.4
trainer                  0.0.36
traitlets                5.14.3
transformers             4.41.0
triton                   2.3.0
TTS                      0.22.0
typeguard                4.2.1
typer                    0.9.4
typing_extensions        4.11.0
tzlocal                  5.2
umap-learn               0.5.6
Unidecode                1.3.8
uritemplate              4.1.1
urllib3                  2.2.1
vosk                     0.3.45
wasabi                   1.1.2
wcwidth                  0.2.13
weasel                   0.3.4
websockets               12.0
Werkzeug                 3.0.3
wheel                    0.43.0
yarl                     1.9.4
zipp                     3.17.0
